Java Fundamentals/src/com/revature/hw1/q22/Q22Test.java
package com.revature.hw1.q22;

import static org.junit.Assert.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

public class Q22Test {

	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void test() {
		int a = 2;
		int b = 3;
		assertEquals(6, (int) Q22.getRecArea(a, b));
		assertEquals(3, (int) Q22.getTriArea(a, b));
	}

}
